doc: mention 'git show' defaults to HEAD
When 'git show' is called without any object it defaults to HEAD.  This
has been true since f2a7a6e0cd ("Simplify common default options setup
for built-in log family.", 2006-04-16).

The SYNOPSIS suggests that the object argument is required.  Clarify
that it is not required and note the default.

Signed-off-by: Todd Zullinger <tmz@pobox.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>