wildmatch test: remove dead fnmatch() test code
Remove the unused fnmatch() test parameter from the wildmatch
test. The code that used to test this was removed in 45882f42bb ("stop
using fnmatch (either native or compat)", 2014-02-15).

As a --word-diff shows the only change to the body of the tests is the
removal of the second out of four parameters passed to match().

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>